Initial version
This source code has been written in GOLANG to allow administration of system users on an ARM v7 based device running Ubuntu and using REDIS as the data storage. It supports a simple structure and allows for adding new users, changing email and updating passwords as well as logging in via a simple JSON API.  The security is provided by the PBKDF2 function using the SHA-256 hash algorithm and a secure randomly generated salt with 4096 iterations to hash the password before storing it.

This is my 1st major project in GO and I am using it to provide a web based user interface that runs on the device to support configuration and maintenance of the software that runs on it. Its reasonably generic and could be of use to anyone who needs to do a similar thing.